% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSprite.checkGrim.R
\name{rSprite.checkGrim}
\alias{rSprite.checkGrim}
\title{Check GRIM}
\usage{
rSprite.checkGrim(N, tMean, dp, rSprite.dust = sqrt(.Machine$double.eps))
}
\arguments{
\item{N}{Number of observations}

\item{tMean}{Reported mean}

\item{dp}{Number of decimal places}

\item{rSprite.dust}{Very small number}
}
\description{
See if a mean is GRIM-consistent. If not, return the nearest mean that is.
}
